==> examples/publication/BMS_yield_cost/0_data_preprocessing.ipynb <==
import numpy as np
import pandas as pd

df = pd.read_csv('./data/PCI_PMI_cost_full_update.csv')
df.columns

for i in range(0, len(df.columns)):
    if len(np.unique(df[df.columns[i]])) > 1:
        print(df.columns[i])

# np.unique(df['Concentration'].values)

df_sel = df[['Base', 'Ligand', 'Solvent', 'Concentration', 'Temp_C', 'Yield', 'total_cost_update', 'PMI']]

df_all_exp_index = df_sel.rename(columns={'Base': 'base', 'Solvent': 'solvent',
                         'Ligand': 'ligand', 'Concentration': 'concentration',
                         'Temp_C': 'temperature', 'Yield': 'yield',
                         'total_cost_update': 'cost'
                        })

# df_all_exp_index.to_csv('./data/experiment_index.csv', index=False)
df_all_exp_index

df_yield_cost = df_all_exp_index.drop(columns=['PMI'])
# df_yield_cost['new_index'] = np.arange(0, len(df_yield_cost))
df_yield_cost.to_csv('./data/experiments_yield_and_cost.csv')



df = pd.read_csv('./data/experiments_yield_and_cost.csv')
df.rename(columns={'Unnamed: 0': 'new_index'}, inplace=True)
df.to_csv('./data/experiments_yield_and_cost.csv', index=False)
